backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
python-multipart==0.0.12
boto3==1.35.0
pillow==10.4.0
jinja2==3.1.4
python-dateutil==2.9.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
